Removing \r from files and moves protocol to `__init__.py`
